Circle fixes: token locations
