+++ src/devtools/client/debugger/src/utils/editor/get-token-location.test.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
-import { getTokenLocation } from "./get-token-location";
+import { getTokenLocation } from "./token-events";
 
 describe("getTokenLocation", () => {
   const codemirror = {
